Creat Melding-package and move files
