--- src/main/kotlin/no/nav/syfo/behandlerdialog/api/MeldingTilBehandlerResponseDTO.kt
@@ -1,11 +1,0 @@
-package no.nav.syfo.behandlerdialog.api
-
-import java.time.OffsetDateTime
-import java.util.*
-
-// Denne vil etterhvert være en slags gruppering på samtaleRef->meldinger
-data class MeldingTilBehandlerResponseDTO(
-    val behandlerRef: UUID,
-    val tekst: String,
-    val bestiltTidspunkt: OffsetDateTime,
-)
