--- src/main/kotlin/no/nav/syfo/behandlerdialog/kafka/DialogmeldingBestillingDTO.kt
@@ -1,13 +1,0 @@
-package no.nav.syfo.behandlerdialog.kafka
-
-data class DialogmeldingBestillingDTO(
-    val behandlerRef: String,
-    val personIdent: String,
-    val dialogmeldingUuid: String,
-    val dialogmeldingRefParent: String?,
-    val dialogmeldingRefConversation: String,
-    val dialogmeldingType: String,
-    val dialogmeldingKode: Int,
-    val dialogmeldingTekst: String?,
-    val dialogmeldingVedlegg: ByteArray? = null,
-)
