--- src/main/kotlin/no/nav/syfo/dialogmelding/domain/DialogmeldingType.kt
@@ -1,7 +1,0 @@
-package no.nav.syfo.dialogmelding.domain
-
-enum class DialogmeldingType() {
-    DIALOG_FORESPORSEL,
-    DIALOG_NOTAT,
-    DIALOG_SVAR,
-}
