--- src/main/kotlin/no/nav/syfo/dialogmelding/kafka/JacksonKafkaDeserializer.kt
@@ -1,13 +1,0 @@
-package no.nav.syfo.dialogmelding.kafka
-
-import no.nav.syfo.dialogmelding.kafka.domain.KafkaDialogmeldingFromBehandlerDTO
-import no.nav.syfo.util.configuredJacksonMapper
-import org.apache.kafka.common.serialization.Deserializer
-
-val mapper = configuredJacksonMapper()
-
-class JacksonKafkaDeserializerDialogmeldingFromBehandler : Deserializer<KafkaDialogmeldingFromBehandlerDTO> {
-    override fun deserialize(topic: String, data: ByteArray): KafkaDialogmeldingFromBehandlerDTO =
-        mapper.readValue(data, KafkaDialogmeldingFromBehandlerDTO::class.java)
-    override fun close() {}
-}
