--- src/main/kotlin/no/nav/syfo/dialogmelding/kafka/KafkaDialogmeldingFraBehandlerTask.kt
@@ -1,22 +1,0 @@
-package no.nav.syfo.dialogmelding.kafka
-
-import no.nav.syfo.application.ApplicationState
-import no.nav.syfo.application.database.DatabaseInterface
-import no.nav.syfo.application.kafka.KafkaEnvironment
-import no.nav.syfo.application.launchBackgroundTask
-
-fun launchKafkaTaskDialogmeldingFraBehandler(
-    applicationState: ApplicationState,
-    kafkaEnvironment: KafkaEnvironment,
-    database: DatabaseInterface,
-) {
-    launchBackgroundTask(
-        applicationState = applicationState,
-    ) {
-        blockingApplicationLogicDialogmeldingFraBehandler(
-            applicationState = applicationState,
-            kafkaEnvironment = kafkaEnvironment,
-            database = database,
-        )
-    }
-}
