+++ src/main/kotlin/no/nav/syfo/melding/domain/DialogmeldingType.kt
@@ -1,0 +1,7 @@
+package no.nav.syfo.melding.domain
+
+enum class DialogmeldingType() {
+    DIALOG_FORESPORSEL,
+    DIALOG_NOTAT,
+    DIALOG_SVAR,
+}
