+++ src/main/kotlin/no/nav/syfo/melding/kafka/config/KafkaBehandlerDialogmeldingSerializer.kt
@@ -1,0 +1,11 @@
+package no.nav.syfo.melding.kafka.config
+
+import no.nav.syfo.melding.kafka.domain.DialogmeldingBestillingDTO
+import no.nav.syfo.util.configuredJacksonMapper
+import org.apache.kafka.common.serialization.Serializer
+
+class KafkaBehandlerDialogmeldingSerializer : Serializer<DialogmeldingBestillingDTO> {
+    private val mapper = configuredJacksonMapper()
+    override fun serialize(topic: String?, data: DialogmeldingBestillingDTO?): ByteArray =
+        mapper.writeValueAsBytes(data)
+}
