+++ src/main/kotlin/no/nav/syfo/melding/kafka/domain/DialogmeldingBestillingDTO.kt
@@ -1,0 +1,13 @@
+package no.nav.syfo.melding.kafka.domain
+
+data class DialogmeldingBestillingDTO(
+    val behandlerRef: String,
+    val personIdent: String,
+    val dialogmeldingUuid: String,
+    val dialogmeldingRefParent: String?,
+    val dialogmeldingRefConversation: String,
+    val dialogmeldingType: String,
+    val dialogmeldingKode: Int,
+    val dialogmeldingTekst: String?,
+    val dialogmeldingVedlegg: ByteArray? = null,
+)
