+++ src/test/kotlin/no/nav/syfo/testhelper/generator/MeldingTilBehandlerGenerator.kt
@@ -1,6 +1,6 @@
 package no.nav.syfo.testhelper.generator
 
-import no.nav.syfo.behandlerdialog.api.MeldingTilBehandlerRequestDTO
+import no.nav.syfo.melding.api.MeldingTilBehandlerRequestDTO
 import java.util.UUID
 
 fun generateMeldingTilBehandlerRequestDTO(
